Move mypyc.ops_str to mypyc.primitives.str_ops
