+++ mypyc/primitives/str_ops.py
